#143 Extract plugin manager components to separate files
