+++ feature/plugin-manager/src/main/java/io/filmtime/feature/plugin/manager/PluginManagerScreen.kt
@@ -6,43 +6,26 @@ import android.net.Uri
 import android.provider.Settings
 import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
-import androidx.compose.material.icons.filled.CheckCircle
-import androidx.compose.material.icons.filled.Extension
-import androidx.compose.material.icons.filled.MoreVert
-import androidx.compose.material3.Card
-import androidx.compose.material3.DropdownMenu
-import androidx.compose.material3.DropdownMenuItem
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.FilledTonalButton
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedButton
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
@@ -51,11 +34,11 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.compose.LifecycleEventEffect
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import io.filmtime.core.designsystem.theme.PreviewFilmTimeTheme
-import io.filmtime.core.designsystem.theme.ThemePreviews
-import io.filmtime.core.plugin.api.PluginAuthState
+import io.filmtime.core.designsystem.composable.FilmTimeCircularProgressBar
 import io.filmtime.core.plugin.api.PluginContract
 import io.filmtime.core.plugin.api.PluginMetadata
+import io.filmtime.feature.plugin.manager.components.EmptyPluginsMessage
+import io.filmtime.feature.plugin.manager.components.PluginCard
 
 @Composable
 fun PluginManagerScreen(
@@ -156,9 +139,23 @@ private fun PluginManagerContent(
       .padding(contentPadding),
     contentPadding = PaddingValues(16.dp),
   ) {
-    if (state.plugins.isEmpty() && !state.isLoading) {
+    if (state.isLoading) {
       item {
-        EmptyPluginsMessage()
+        FilmTimeCircularProgressBar(
+          modifier = Modifier
+            .animateItem()
+            .fillParentMaxSize()
+            .wrapContentSize(),
+        )
+      }
+    } else if (state.plugins.isEmpty()) {
+      item {
+        EmptyPluginsMessage(
+          modifier = Modifier
+            .animateItem()
+            .fillParentMaxSize()
+            .wrapContentSize(),
+        )
       }
     } else {
       items(state.plugins, key = { it.pluginId }) { plugin ->
@@ -184,171 +181,5 @@ private fun PluginManagerContent(
         )
       }
     }
-  }
-}
-
-@Composable
-private fun PluginCard(
-  plugin: PluginMetadata,
-  isDefault: Boolean,
-  authState: PluginAuthState?,
-  onClick: () -> Unit,
-  onLoginClick: () -> Unit,
-  onLogoutClick: () -> Unit,
-  onViewInfoClick: () -> Unit,
-  modifier: Modifier = Modifier,
-) {
-  var menuExpanded by remember { mutableStateOf(false) }
-
-  Card(
-    modifier = modifier
-      .fillMaxWidth()
-      .clickable(onClick = onClick),
-  ) {
-    Row(
-      modifier = Modifier
-        .fillMaxWidth()
-        .padding(start = 16.dp)
-        .padding(vertical = 16.dp),
-      verticalAlignment = Alignment.CenterVertically,
-    ) {
-      Icon(
-        modifier = Modifier
-          .padding(top = 4.dp)
-          .align(Alignment.Top),
-        imageVector = Icons.Default.Extension,
-        contentDescription = null,
-        tint = MaterialTheme.colorScheme.primary,
-      )
-      Spacer(modifier = Modifier.width(16.dp))
-      Column(
-        modifier = Modifier.weight(1f),
-      ) {
-        Text(
-          text = plugin.name,
-          style = MaterialTheme.typography.titleMedium,
-        )
-        Text(
-          text = plugin.description,
-          style = MaterialTheme.typography.bodyMedium,
-          color = MaterialTheme.colorScheme.onSurfaceVariant,
-        )
-        Text(
-          text = stringResource(R.string.plugin_manager_version, plugin.version),
-          style = MaterialTheme.typography.bodySmall,
-          color = MaterialTheme.colorScheme.onSurfaceVariant,
-        )
-        if (plugin.requiresAuth && authState != null) {
-          Spacer(modifier = Modifier.height(8.dp))
-          when (authState) {
-            is PluginAuthState.Authenticated -> {
-              OutlinedButton(onClick = onLogoutClick) {
-                Text(stringResource(R.string.plugin_manager_logout))
-              }
-            }
-            is PluginAuthState.NotAuthenticated -> {
-              FilledTonalButton(onClick = onLoginClick) {
-                Text(stringResource(R.string.plugin_manager_login))
-              }
-            }
-            else -> {}
-          }
-        }
-      }
-      if (isDefault) {
-        Icon(
-          imageVector = Icons.Default.CheckCircle,
-          contentDescription = stringResource(R.string.plugin_manager_default),
-          tint = MaterialTheme.colorScheme.primary,
-        )
-      }
-      Box {
-        IconButton(onClick = { menuExpanded = true }) {
-          Icon(
-            imageVector = Icons.Default.MoreVert,
-            contentDescription = stringResource(R.string.plugin_manager_menu),
-          )
-        }
-        DropdownMenu(
-          expanded = menuExpanded,
-          onDismissRequest = { menuExpanded = false },
-        ) {
-          DropdownMenuItem(
-            text = {
-              Text(
-                stringResource(
-                  if (isDefault) {
-                    R.string.plugin_manager_remove_default
-                  } else {
-                    R.string.plugin_manager_set_default
-                  },
-                ),
-              )
-            },
-            onClick = {
-              menuExpanded = false
-              onClick()
-            },
-          )
-          DropdownMenuItem(
-            text = { Text(stringResource(R.string.plugin_manager_view_info)) },
-            onClick = {
-              menuExpanded = false
-              onViewInfoClick()
-            },
-          )
-        }
-      }
-    }
-  }
-}
-
-@Composable
-private fun EmptyPluginsMessage() {
-  Column(
-    modifier = Modifier
-      .fillMaxWidth()
-      .padding(32.dp),
-    horizontalAlignment = Alignment.CenterHorizontally,
-  ) {
-    Icon(
-      imageVector = Icons.Default.Extension,
-      contentDescription = null,
-      tint = MaterialTheme.colorScheme.onSurfaceVariant,
-    )
-    Spacer(modifier = Modifier.height(16.dp))
-    Text(
-      text = stringResource(R.string.plugin_manager_no_plugins_title),
-      style = MaterialTheme.typography.titleMedium,
-    )
-    Spacer(modifier = Modifier.height(8.dp))
-    Text(
-      text = stringResource(R.string.plugin_manager_no_plugins_description),
-      style = MaterialTheme.typography.bodyMedium,
-      color = MaterialTheme.colorScheme.onSurfaceVariant,
-    )
-  }
-}
-
-@ThemePreviews
-@Composable
-private fun PluginCardPreview() {
-  PreviewFilmTimeTheme {
-    PluginCard(
-      plugin = PluginMetadata(
-        pluginId = "plugin-id",
-        name = "Test Plugin",
-        description = "This is a test plugin",
-        version = "1.0.0",
-        iconUrl = null,
-        authority = "io.filmtime.test.plugin",
-      ),
-      isDefault = true,
-      authState = PluginAuthState.NotAuthenticated,
-      onClick = {},
-      onLoginClick = {},
-      onLogoutClick = {},
-      onViewInfoClick = {},
-    )
   }
 }
