+++ feature/plugin-manager/src/main/java/io/filmtime/feature/plugin/manager/components/EmptyPluginsMessage.kt
@@ -1,0 +1,62 @@
+package io.filmtime.feature.plugin.manager.components
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Extension
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import io.filmtime.core.designsystem.theme.PreviewFilmTimeTheme
+import io.filmtime.core.designsystem.theme.ThemePreviews
+import io.filmtime.feature.plugin.manager.R
+
+@Composable
+internal fun EmptyPluginsMessage(
+  modifier: Modifier = Modifier,
+) {
+  Column(
+    modifier = modifier
+      .fillMaxWidth()
+      .padding(16.dp),
+    horizontalAlignment = Alignment.CenterHorizontally,
+  ) {
+    Icon(
+      modifier = Modifier
+        .size(32.dp),
+      imageVector = Icons.Default.Extension,
+      contentDescription = null,
+      tint = MaterialTheme.colorScheme.onSurfaceVariant,
+    )
+    Spacer(modifier = Modifier.height(16.dp))
+    Text(
+      text = stringResource(R.string.plugin_manager_no_plugins_title),
+      style = MaterialTheme.typography.titleMedium,
+    )
+    Spacer(modifier = Modifier.height(8.dp))
+    Text(
+      text = stringResource(R.string.plugin_manager_no_plugins_description),
+      style = MaterialTheme.typography.bodyMedium,
+      color = MaterialTheme.colorScheme.onSurfaceVariant,
+      textAlign = TextAlign.Center,
+    )
+  }
+}
+
+@ThemePreviews
+@Composable
+private fun EmptyPluginsMessagePreview() {
+  PreviewFilmTimeTheme {
+    EmptyPluginsMessage()
+  }
+}
